+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## Unreleased
 
+# Added
+
+- Add --generate-temp-token CLI argument to rt websocket setup to get temp token for rt authentication
+- Add generate_temp_token optional boolean kwarg to connection setttings, defaults to False
+
 ## [1.3.0] - 2022-08-05
 
 # Added